feat: Update secureboot.nix to include reboot-to-windows package and dependencies
This commit updates the `secureboot.nix` file to include the `reboot-to-windows` package and its dependencies. The `reboot-to-windows` package allows users to change the EFI boot order and reboot the system to Windows. It includes a shell script for modifying the boot order and a desktop item for easy access. The package dependencies include `efibootmgr` and `makeDesktopItem`. This update enhances the functionality of the `secureboot` module for Surface devices.

Co-authored-by: timoverbrugghe