feat: Add extraUpFlags to Tailscale configuration
This commit adds the `extraUpFlags` option to the Tailscale configuration in the Surface configuration file. The `extraUpFlags` allow for additional command-line flags to be passed to the Tailscale service, in this case, the `--operator` flag with the value of the username. This change enhances the functionality and customization options of the Tailscale service for the Surface device.

Co-authored-by: timoverbrugghe