refactor: Update secureboot.nix to fix path for reboot-to-windows package
This commit updates the secureboot.nix file in the nixos/machines/surface/secureboot directory to fix the path for the reboot-to-windows package. The previous path was incorrect, causing the package to not be included in the build. This change ensures that the package is correctly included and improves the functionality of the secureboot module for Surface devices.

Co-authored-by: timoverbrugghe