refactor: Update chrome app launch commands in package.nix
This commit updates the launch commands for the Chrome apps in the package.nix file. The previous commands were using the absolute path to the Google Chrome executable, which is not necessary. By removing the absolute path, the launch commands will use the default system installation of Google Chrome. This change simplifies the configuration and ensures compatibility with different system setups.

Co-authored-by: timoverbrugghe